+++ src/SGDE.Domain/Entities/DetailInvoice.cs
@@ -9,9 +9,9 @@
     public class DetailInvoice : BaseEntity
     {
         public string ServicesPerformed { get; set; }
-        public Decimal Units { get; set; }
-        public Decimal UnitsAccumulated { get; set; }
-        public Decimal PriceUnity { get; set; }
+        public double Units { get; set; }
+        public double UnitsAccumulated { get; set; }
+        public double PriceUnity { get; set; }
         public string NameUnit { get; set; }
         public double Total => (double)Units * (double)PriceUnity;
