+++ src/SGDE.Domain/Supervisor/SupervisorWork.cs
@@ -79,9 +79,8 @@
                 WorksToRealize = newWorkViewModel.worksToRealize,
                 Open = newWorkViewModel.open,
                 InvoiceToOrigin = newWorkViewModel.invoiceToOrigin,
-                //TotalContract = (decimal)newWorkViewModel.totalContract,
-                PercentageRetention = (decimal)newWorkViewModel.percentageRetention,
-                PercentageIVA = (decimal)newWorkViewModel.percentageIVA,
+                PercentageRetention = newWorkViewModel.percentageRetention,
+                PercentageIVA = newWorkViewModel.percentageIVA,
 
                 OpenDate = newWorkViewModel.openDate == null ? DateTime.Now : DateTime.Parse(newWorkViewModel.openDate),
                 CloseDate = string.IsNullOrEmpty(newWorkViewModel.closeDate)
@@ -113,9 +112,8 @@
             work.EstimatedDuration = workViewModel.estimatedDuration;
             work.WorksToRealize = workViewModel.worksToRealize;
             work.InvoiceToOrigin = workViewModel.invoiceToOrigin;
-            //work.TotalContract = (decimal)workViewModel.totalContract;
-            work.PercentageRetention = (decimal)workViewModel.percentageRetention;
-            work.PercentageIVA = (decimal)workViewModel.percentageIVA;
+            work.PercentageRetention = workViewModel.percentageRetention;
+            work.PercentageIVA = workViewModel.percentageIVA;
 
             work.OpenDate = DateTime.ParseExact(workViewModel.openDate, "dd/MM/yyyy", null);
             work.CloseDate = string.IsNullOrEmpty(workViewModel.closeDate)
